Change type of uspace pointers in kernel from pointer type to numeric type
From kernel's perspective, userspace addresses are not valid pointers,
and can only be used in calls to `copy_to/from_uspace()`.
Therefore, we change the type of those arguments and variables to
`uspace_addr_t` which is an alias for `sysarg_t`.

This allows the compiler to catch accidental direct accesses to
userspace addresses.

Additionally, to avoid losing the type information in code,
a macro `uspace_ptr(type)` is used that translates to `uspace_addr_t`.
I makes no functional difference, but allows keeping the type information
in code in case we implement some sort of static checking for it in the future.

+++ kernel/arch/arm32/src/userspace.c
@@ -70,7 +70,7 @@ void userspace(uspace_arg_t *kernel_uarg)
 	volatile ustate_t ustate;
 
 	/* set first parameter */
-	ustate.r0 = (uintptr_t) kernel_uarg->uspace_uarg;
+	ustate.r0 = kernel_uarg->uspace_uarg;
 
 	/* %r1 is defined to hold pcb_ptr - set it to 0 */
 	ustate.r1 = 0;
@@ -92,11 +92,11 @@ void userspace(uspace_arg_t *kernel_uarg)
 	ustate.lr = 0;
 
 	/* set user stack */
-	ustate.sp = ((uint32_t) kernel_uarg->uspace_stack) +
+	ustate.sp = kernel_uarg->uspace_stack +
 	    kernel_uarg->uspace_stack_size;
 
 	/* set where uspace execution starts */
-	ustate.pc = (uintptr_t) kernel_uarg->uspace_entry;
+	ustate.pc = kernel_uarg->uspace_entry;
 
 	/* status register in user mode */
 	ipl_t user_mode = current_status_reg_read() &
